Inclusion of flex file to implement file scanner
Currently very limited. After file is scanned data is saved into a struct containing the number of columns, the number of rows, a char** containing the title of each column, and a char*** containing the data of each row of each column. future version will need some way of enforcing data types on inputs from the file